service/requirement.txt
pytz==2024.1
requests==2.31.0
beautifulsoup4==4.12.3
PyYAML==6.0.1